Checkpoint - splitting up effects headers
