+++ include/effects/strip/bouncingballeffect.h
