+++ include/effects/strip/doublepaletteeffect.h
