+++ include/effects/strip/faneffects.h
