+++ include/effects/strip/fireeffect.h
