+++ include/effects/strip/meteoreffect.h
