+++ include/effects/strip/misceffects.h
