+++ include/effects/strip/musiceffect.h
