+++ include/effects/strip/paletteeffect.h
