+++ include/effects/strip/particles.h
