+++ include/effects/strip/stareffect.h
