+++ include/effects/strip/tempeffect.h
