+++ include/ledstripeffect.h
