+++ src/screen.cpp
@@ -31,7 +31,7 @@
 #include "globals.h"                            // CONFIG and global headers
 #include "gfxbase.h"
 #include "ledbuffer.h"                          // For g_apBufferManager type
-#include "effects/effectmanager.h"              // So we can display cur effect
+#include "effectmanager.h"              // So we can display cur effect
 #include "Bounce2.h"
 #include "freefonts.h"
 #include "colordata.h"
